% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateCohorts.R
\name{generateAllCohorts}
\alias{generateAllCohorts}
\title{Genarates all the cohorts of the LEGEND-hyperetnsion study}
\usage{
generateAllCohorts(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  oracleTempSchema = NULL,
  indicationId = "Hypertension",
  outputFolder
)
}
\description{
Genarates all the cohorts of the LEGEND-hyperetnsion study
}
